requirements.txt
faker
pymongo
random
logging
flask
tkinter
request
subprocess
time